+++ common/src/main/java/club/iananderson/seasonhud/client/gui/screens/SeasonHudScreen.java
@@ -84,7 +84,7 @@ public class SeasonHudScreen extends Screen {
     this.rebuildWidgets();
   }
 
-  public void drawButtonRow(){
+  public void drawButtonRow() {
 
   }