+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Changed
 
 - (1.20.1 and Below) Changed Journeymap implementation to try and better support Retina displays on macOS
-- Switched back to Ecliptic Season's translations for its sub-seasons
+- (Ecliptic Seasons) Switched back to Ecliptic Seasons' translations for its sub-seasons
 
 ## [2.0.0] - 2026-02-27